added some global defaults
Enabled global and module specific traceDelays in addition to the per channel ones.

Enabled Global Firmware and Frequency for nonmixed setups.
I also changed the what the defaultChannelParms are set so that even if the Trace and Fit sub nodes arent there AND its for a vandle it sets the defaults since they are for vandle:medium anyways