Big Commit of Root Stuff and new Processor
systemwide root output, and the new GammaScintProcessor.  Currently the Clover, Vandle, and GammaScint Processors support root output.

Also added the pspmt dynode as a valid start for the vandle processor. i used the new types/subtypes for the pspmt processor (which will be coming in a later commit after further testing)