Updates to POLL2 messages
Updated 'csr_test' to included bit 22, the external timing one.

Increased the Digits for the preads so that we we stay with full ints when the Ext Timing bit is on.

I also swapped the "HI" and "LO" order in the help message to better reflect that LOw Gain is when the bit is 0 (off)