Updated GammaScint to take the HAGRiD dynodes
Updated the code to not include the dynodes into things like the totals BG spectra, and the addback calculations. I added a simple dynode totals plot so that we can watch it. I also updated the root structure so that we have distinction between the normal outputs and the dynodes. We will need to add the "dy" tag to the dynode channels and it should be noted, since we have not decided on a naming convention for the addback groups, that  the group tag MUST be first in the tag list.