Fixed the Onboard CFD decoding and added the mask for R34300
R34300 firmware uses the same word 2 structure as the newer R34688 so we just duplicated the masks and bit shifts.

The CfdTriggerSourceBit was not being decoded correctly due to a missing () around the Boolean cast.

Notes about changes needed for the new revF 500Mhz modules. And a typo fix here and there