Added option for the generation of the compile_commands.json
I had to wrap the actual command in a new PAASS_ wrapper because "FORCE" is required for the variable to be set, however this removes the ability to turn it off. 

I compared ON and OFF on my laptop (ideapad 730s i7 16GB) and the flag on only added ~0.5s on to (cmake ../ -DPAASS_BUILD_UTKSCAN=ON -DPAASS_USE_HRIBF=ON  && make -j7; ) which i believe to be within the margins since the computer was in use during both runs.