New Mixed Crates Tick to TIme conversions
Added the module frequency to the channel configuration class, needed so that we have access to the module frequency for each specific channel in the crate.

New Overload to the GetClockInSeconds() family. These take the module frequency as an input and return the correct conversion factor. This will allow for mixed frequency crates to correctly convert ticks to times.
For legacy/compatibility reasons, if no argument is passed  we default  to 250MHz as this is our most common type.  Users of mixed crates are expected to pass the frequency; and this should be the default behavior going forward.

The RootDevProcessor.cpp has an example of how to utilize these new methods.