Generalized the PSPMT dynode trace analysis
Rin moved the RIKEN (ribf136) specific code here, we have generalized it as much as possible, leveraging what paass is already doing. As a result we have some new fields in the pspmt root output.  They should be self explanatory, but here is a run through.
traceMaxPos and traceMaxVal are the location of the maximum in the trace and its value, the preBaseAvg is the baseline average computed in the waveform analyzer, the postBaseAvg is the average of the last 20 bins in the trace, and the invalidTrace bool is true when there is large difference between the min and max values in the pre waveform baseline (set to 40 right now) . It is also true for channels without traces.

Additionally there was some formatting fixes to the PspmtProcessor.cpp file (mostly whitespace, and indentation fixes)