updated VandleProcessor for Onboard CFD starts
added support for the SingleBetaProcessor starts (normal trace fitting as well as OCFD)

as well as general support for the OnboardCFD to be used in the toc calculation. 

also updated the root output field vandles.sTime  to output in ns like everyone else. and to protect the sQdc field by the new hasValidWaveFormAnalysis.\

also made a note in the DetectorDriver that we should update GetSummary() to take "type:subtype:group:tag" combinations. Right now it only does type:subtype and the start tag.