DAMM Histogram Declaration Suppression
We can now suppress the declaration of histogram, using the same flag as the last commit. For UTKSCAN, we can Fully suppress the declaration, leading to a .his of size 0. but for UTKSCANOR we need to have atleast 1 histogram to pass a sanity check somewhere in the scanor fortran code. So the DetectorDriver will create a simple 1d for this purpose. It will only create it if we build UTKSCANOR and have the DAMM plotting turned off.