Reimplemented the poll2 email alarm
poll2 can send an alarm email when it enters an ERROR state.
The syntax is `poll2 -a<email>` with no space between the `-a` and the email address or `poll2 --alarm=<email>` (no space between the = and the email)
the emails here can be a CSV list of emails so multiple people can receive the alerts.
If there is expressed interest we can try to implement the functionality to read the emails from the `pixie.cfg` file. But this would not be  preferred.
We also "install" the send_alarm bash script so that its in PATH when we try to call it.
This functionality also requires Users to setup the `mail` command on the DAQ computer. See the wiki for basic instructions.