Changes to extreme baseline variation
Changing the extreme baseline threshold to 15% of the baseline value, We are doing this because 16b can have a variation of 300-400 and still be ok. 

Im also adding an error statement to this, so that for every 10,000 traces we throw out here; we print a note into the terminal. 

Also adding a cut for the avg baseline also being <10 adc units (for every bit resolution) because if the base line is this low then something is wrong. (example: beta traces for RIKEN after implant)