Changing Ctrl+c in poll2 to clear the cmd line
This commit changes  the use of <Ctrl>+c in poll2.
Previously all we did was catch the <ctrl>+c and print an error. Now we are using this key stroke to clear the current line in the poll2 prompt. It does not store the partial command in the history.

Currently we are just wrapping the private function Terminal::clear_() with a publicly accessible version, but we should probably reimplement the clearing action with the public nature in-mind

 This has a limited use case but is helpful on occasion.