Updating firmware support
We now fully support the 250MHz firmware received in  2019, namely 42159 and 42950 which are the newest 12b and 16b vandle firmwares respectively

We also have initial support for the 500MHz general firmware. Its hard matched right now while we are confirming everything, but it should have the same list mode as the 34688->42950 series, with the exception of the changes to header word 2 relevant to the 500MHz (changes to the onboard cfd data). If everything checks out we will remove the hard match, and the separate enum value