#97: FindXIA.cmake finds firmware properly. Moved DAQ Config function
FindXIA.cmake wasn't able to find the firmware
directories properly. It was globbing together a couple
directories that didn't make sense.

The DAQ configuration should be separate from the
process of locating XIA's Pixie SDK. I've moved that to
its own file and organized it under Cmake/functions.
This gives us the ability in the future to generalize it
to any company's libraries.