#97: Reorganizing files to remove circular dependencies
Some files didn't live in reasonable places. We needed
to move the Lock class into Core as that's a more
logical place for it to live. Moved the poll2_socket
stuff to live with poll2. Moved hribf_buffers to a
DataFormats directory, it makes sense to collect all of
these together.